V-02 рефактор хранения компонентов
