--- src/components/elements/Container/Container.module.scss
@@ -1,15 +1,0 @@
-@use '/src/sass/tools' as *;
-
-.container {
-  margin: 0 auto;
-  max-width: 1376px;
-  padding: 0 64px;
-
-  @include media($xl) {
-    padding: 0 32px;
-  }
-
-  @include media($md) {
-    padding: 0 16px;
-  }
-}
