--- src/components/elements/Start/Start.module.scss
@@ -1,6 +1,0 @@
-@use '/src/sass/tools' as *;
-
-.start {
-  @include flex(space-between,_,column);
-  flex: 1 1 auto;
-}
