+++ src/components/elements/common/Container/Container.module.scss
@@ -1,0 +1,15 @@
+@use '../../../../sass/tools' as *;
+
+.container {
+  margin: 0 auto;
+  max-width: 1376px;
+  padding: 0 64px;
+
+  @include media($xl) {
+    padding: 0 32px;
+  }
+
+  @include media($md) {
+    padding: 0 16px;
+  }
+}
