+++ src/components/elements/common/Header/Header.module.scss
@@ -1,0 +1,40 @@
+@use '../../../../sass/tools' as *;
+
+.header {
+  h1 {
+    line-height: 35.16px;
+  }
+}
+
+.body {
+  padding: 32px 0;
+  @include flex(space-between,center,_);
+
+  @include media($md) {
+    padding: 16px 0;
+    @include flex(_,flex-end,column);
+  }
+}
+
+.geo {
+  @include flex(_,center,_);
+  font-size: 0.875rem;
+  color: $gray;
+  &:hover {
+    color: $green;
+    svg {
+      path {
+        fill: none;
+        stroke: $green;
+      }
+    }
+  }
+
+  @include media($md) {
+    margin-top: 8px;
+  }
+
+  svg {
+    height: 24px;
+  }
+}
