+++ src/components/elements/homepage/slider-components/Slide/Slide.module.scss
@@ -1,0 +1,50 @@
+@use '/src/sass/tools' as *;
+
+.slide {
+  flex: 0 0 100%;
+  height: 100%;
+  width: 100%;
+  background-size: cover;
+  background-repeat: no-repeat;
+  background-position: center;
+  background-blend-mode: darken;
+}
+
+.body {
+  margin: 28vh 0 0 96px;
+  width: 73%;
+
+  @include media($xl) {
+    margin: 32vh 0 0 64px;
+    width: 64%;
+  }
+}
+
+.title {
+  font-size: 2.5rem;
+  line-height: 48px;
+  font-weight: 500;
+  color: $white;
+
+  @include media($xl) {
+    font-size: 2rem;
+    line-height: 37px;
+  }
+}
+
+.text {
+  margin-top: 8px;
+  font-size: 1.5rem;
+  line-height: 22px;
+  font-weight: 300;
+  color: $gray-light;
+
+  @include media($xl) {
+    font-size: 1.375rem;
+  }
+}
+
+.button {
+  margin-top: 32px;
+  border-radius: 4px;
+}
