+++ src/components/elements/homepage/slider-components/Slider/Slider.module.scss
@@ -1,0 +1,59 @@
+@use '/src/sass/tools' as *;
+
+.slider {
+  position: relative;
+  flex: 0 0 50%;
+  overflow: hidden;
+
+  @include media($xl) {
+    flex: 0 1 40%;
+  }
+
+  @include media($lg) {
+    display: none;
+  }
+}
+
+.content {
+  @include flex (_,_,_);
+  height: 100%;
+  @include transition(0.5s,0s);
+}
+
+.arrow {
+  position: absolute;
+  top: 0%;
+  @include flex(_,center,_);
+  height: 100%;
+  z-index: 3;
+  padding: 0 28px;
+  @include transition(0.2s,0s);
+  cursor: pointer;
+  svg {
+    path {
+      fill: none;
+    }
+  }
+
+  &:hover {
+    background-color: rgba($green,0.2);
+    @include transition(0.2s,0s);
+  }
+
+  &:hover {
+    svg {
+      opacity: 1;
+    }
+  }
+
+  &_left {
+    left: 0;
+  }
+  &_right {
+    right: 0;
+  }
+
+  @include media($xl) {
+    padding: 0 12px;
+  }
+}
