+++ src/components/elements/menu/LangSwitcher/LangSwitcher.module.scss
@@ -1,0 +1,29 @@
+@use '../../../../sass/tools' as *;
+
+.switcher {
+  @include flex(center,center,_);
+  width: 48px;
+  height: 48px;
+  font-weight: 700;
+  font-size: 0.8125rem;
+  line-height: 15px;
+  color: $green;
+  border-radius: 50%;
+  border: 1px solid transparent;
+  @include transition(0.2s,0s);
+  cursor: pointer;
+
+  &:hover {
+    color: $white;
+    border: 1px solid $gray-light;
+    @include transition(0.2s,0s);
+  }
+  &:active {
+    color: $green;
+    @include transition(0.2s,0s);
+  }
+
+  @include media($md) {
+    display: none;
+  }
+}
