+++ src/components/pages/Homepage/Homepage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Slider from '../../elements/slider-components/Slider/Slider';
+import Slider from '../../elements/homepage/slider-components/Slider/Slider';
 import styles from './Homepage.module.scss';
-import Start from '../../elements/Start/Start';
+import Start from '../../elements/homepage/Start/Start';
 import slides from './slides';
 
 function Homepage() {
